Extract additional UDFs, add TODOs
